refactor(crew): split Testing & Serialization region into CrewReservationManager.Serialization.cs
